server/requirements.txt
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
pandas==2.2.2
paramiko==3.4.0
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sshtunnel==0.4.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
